+++ Toggly.FeatureManagement.Solid/src/server.ts
@@ -116,7 +116,7 @@ export function createTogglyRequest(options: ServerRequestOptions) {
             'User-Agent': context.request?.userAgent ?? '',
             'Accept-Language': context.request?.acceptLanguage ?? '',
             'X-Toggly-Sdk': 'solidstart',
-            'X-Toggly-Sdk-Version': '0.2.0',
+            'X-Toggly-Sdk-Version': '0.3.0',
           },
         },
       );